+++ test/BlocModel.dart
@@ -1,7 +1,7 @@
 import 'package:blocstar/ActionState.dart';
 import 'package:blocstar/BlocModelBase.dart';
 
-class BlocModel extends BlocModelBase<BlocModel> {
+class BlocModel extends BlocContextBase<BlocModel> {
   BlocModel(Function(ActionState) onAppStateChanged) : super(onAppStateChanged);
 
   @override